requirements.txt
Flask==3.0.0
python-frontmatter==1.1.0
markdown2==2.4.12
